Renamed packages to better correlate to module names
