+++ src/backend/mainController/Ambulance.java
@@ -1,4 +1,4 @@
-package backend.simulation;
+package backend.mainController;
 
 public class Ambulance {
     private Point location;
