+++ src/backend/mainController/InjurySeverity.java
@@ -1,4 +1,4 @@
-package backend.simulation;
+package backend.mainController;
 
 public enum InjurySeverity {
     LIFE_THREATENING, NON_LIFE_THREATENING
