+++ src/backend/mainController/Patient.java
@@ -1,4 +1,4 @@
-package backend.simulation;
+package backend.mainController;
