+++ src/backend/mainController/tests/AmbulanceTest.java
@@ -1,7 +1,7 @@
-package backend.simulation.tests;
+package backend.mainController.tests;
 
-import backend.simulation.Ambulance;
-import backend.simulation.Point;
+import backend.mainController.Ambulance;
+import backend.mainController.Point;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
