+++ src/backend/mainController/tests/HospitalTest.java
@@ -1,7 +1,7 @@
-package backend.simulation.tests;
+package backend.mainController.tests;
 
-import backend.simulation.Hospital;
-import backend.simulation.Point;
+import backend.mainController.Hospital;
+import backend.mainController.Point;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
