+++ src/gui/gui.fxml
@@ -1,5 +1,5 @@
 <?import javafx.scene.layout.GridPane?>
 
-<GridPane fx:controller="frontend.GUIController"
+<GridPane fx:controller="gui.GUIController"
           xmlns:fx="http://javafx.com/fxml" alignment="center" hgap="10" vgap="10">
 </GridPane>
